It now looks more like Aternos
You said you wanted it to look more like Aternos so I did my best. Of course is not optimized and I didn't finish it because it's just a template. The styling is now similar to the Aternos team page. Since it's just a test I will not commit it directly but I will make a pull request to see if you like it. If you do, I will optimize it.